Add shops budget and completed/cancelled order registries
