+++ backend/src/main/java/com/tfg/backend/controller/ShopRestController.java
@@ -89,7 +89,7 @@ public class ShopRestController {
     @Operation(summary = "(Admin) Create shop")
     @PostMapping
     public ResponseEntity<ShopDTO> createShop(@RequestBody ShopDTO shopDTO) {
-        Shop savedShop = shopService.createShop(shopDTO);
+        Shop savedShop = shopUserOrchestrator.createShop(shopDTO);
 
         URI location = ServletUriComponentsBuilder
                 .fromCurrentRequest()
@@ -103,7 +103,7 @@ public class ShopRestController {
     @Operation(summary = "(Admin, Manager) Update shop by ID")
     @PutMapping("/{id}")
     public ResponseEntity<ShopDTO> updateShop(@PathVariable Long id, @RequestBody ShopDTO shopDTO) {
-        Shop updatedShop = shopService.updateShop(id, shopDTO);
+        Shop updatedShop = shopUserOrchestrator.updateShop(id, shopDTO);
         return ResponseEntity.accepted().body(toEnrichedDTO(updatedShop));
     }
 
@@ -111,7 +111,7 @@ public class ShopRestController {
     @Operation(summary = "(Admin) Delete shop by ID")
     @DeleteMapping("/{id}")
     public ResponseEntity<ShopDTO> deleteShop(@PathVariable Long id) {
-        Shop deletedShop = shopService.deleteShop(id);
+        Shop deletedShop = shopUserOrchestrator.deleteShop(id);
         return ResponseEntity.ok(toEnrichedDTO(deletedShop));
     }
