+++ backend/src/test/java/com/tfg/backend/unit/ShopServiceUTest.java
@@ -108,96 +108,7 @@ class ShopServiceUTest {
         }
     }
 
-    // --- CRUD TESTS ---
-    @Nested
-    @DisplayName("Tests for basic Shop creation and updates")
-    class CrudTests {
 
-        @Test
-        @DisplayName("createShop maps DTO properties to a new Entity and saves it")
-        void createShop_Success() {
-            when(shopRepository.save(any(Shop.class))).thenAnswer(i -> i.getArgument(0));
-
-            Shop result = shopService.createShop(shopDTO);
-
-            assertEquals("New Shop Name", result.getName());
-            assertEquals(5000.0, result.getAssignedBudget());
-            assertNotNull(result.getAddress());
-            assertEquals("Main Store", result.getAddress().getAlias());
-            assertEquals(40.4168, result.getAddress().getLatitude());
-
-            verify(shopRepository).save(any(Shop.class));
-        }
-
-        @Test
-        @DisplayName("updateShop updates existing entity fields using DTO data")
-        void updateShop_Success() {
-            when(shopRepository.findById(1L)).thenReturn(Optional.of(shop));
-
-            Shop result = shopService.updateShop(1L, shopDTO);
-
-            assertEquals("New Shop Name", result.getName());
-            assertEquals(5000.0, result.getAssignedBudget());
-            assertEquals("Madrid", result.getAddress().getCity());
-        }
-
-        @Test
-        @DisplayName("deleteShop unlinks memory collections, cancels active orders and deletes S3 image")
-        void deleteShop_CascadesAndUnlinks() {
-            // Setup Truck
-            Truck truck = new Truck();
-            truck.setAssignedShop(shop);
-            shop.getAssignedTrucks().add(truck);
-
-            // Setup Customer
-            User customer = new User();
-            customer.setSelectedShop(shop);
-            shop.getCustomers().add(customer);
-
-            // Setup Orders (One active, one completed)
-            Order activeOrder = new Order();
-            activeOrder.changeOrderStatus(OrderStatus.ORDER_MADE, "Init");
-            activeOrder.setAssignedShop(shop);
-
-            Order completedOrder = new Order();
-            completedOrder.changeOrderStatus(OrderStatus.COMPLETED, "Done");
-            completedOrder.setAssignedShop(shop);
-
-            shop.getAssignedOrders().addAll(List.of(activeOrder, completedOrder));
-
-            when(shopRepository.findById(1L)).thenReturn(Optional.of(shop));
-
-            try (MockedStatic<GlobalDefaults> mockedDefaults = mockStatic(GlobalDefaults.class)) {
-                mockedDefaults.when(() -> GlobalDefaults.isDefaultShopImage(shop.getImage())).thenReturn(false);
-
-                // Act
-                shopService.deleteShop(1L);
-
-                // Assert unlinking
-                assertTrue(shop.getAssignedTrucks().isEmpty());
-                assertNull(truck.getAssignedShop());
-
-                assertTrue(shop.getCustomers().isEmpty());
-                assertNull(customer.getSelectedShop());
-
-                assertTrue(shop.getAssignedOrders().isEmpty());
-                assertNull(activeOrder.getAssignedShop());
-                assertNull(completedOrder.getAssignedShop());
-
-                // Assert business logic: Active order must be cancelled
-                assertEquals(OrderStatus.CANCELLED, activeOrder.getCurrentStatus());
-                assertEquals("La tienda a la que estaba asignado el pedido ha sido eliminada.",
-                        activeOrder.getHistory().getLast().getUpdates().getLast().getDescription());
-
-                // Completed order must remain completed
-                assertEquals(OrderStatus.COMPLETED, completedOrder.getCurrentStatus());
-
-                // Assert deletions
-                verify(imageService).deleteFile("shop-pic.jpg");
-                verify(shopRepository).delete(shop);
-            }
-        }
-    }
 
     // --- STOCK MANAGEMENT TESTS ---
     @Nested
