+++ backend/src/test/java/com/tfg/backend/unit/ShopUserOrchestratorUTest.java
@@ -2,9 +2,17 @@ package com.tfg.backend.unit;
 
 import com.tfg.backend.model.Shop;
 import com.tfg.backend.model.User;
+import com.tfg.backend.model.Truck;
+import com.tfg.backend.model.Order;
+import com.tfg.backend.model.OrderStatus;
+import com.tfg.backend.model.ImageInfo;
+import com.tfg.backend.dto.ShopDTO;
+import com.tfg.backend.dto.AddressDTO;
 import com.tfg.backend.service.ShopService;
 import com.tfg.backend.service.ShopUserOrchestrator;
 import com.tfg.backend.service.UserService;
+import com.tfg.backend.service.ImageService;
+import com.tfg.backend.utils.GlobalDefaults;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
@@ -12,6 +20,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
+import org.mockito.MockedStatic;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.data.domain.Page;
@@ -32,6 +41,7 @@ class ShopUserOrchestratorUTest {
 
     @Mock private ShopService shopService;
     @Mock private UserService userService;
+    @Mock private ImageService imageService;
     @Mock private ApplicationEventPublisher eventPublisher; //Necessary to avoid errors trying to send notifications, but not used
 
     @InjectMocks
@@ -40,6 +50,8 @@ class ShopUserOrchestratorUTest {
     private User loggedUser;
     private Shop shop;
     private User managerUser;
+    private ShopDTO shopDTO;
+    private AddressDTO addressDTO;
 
     @BeforeEach
     void setUp() {
@@ -52,6 +64,113 @@ class ShopUserOrchestratorUTest {
         shop = new Shop();
         shop.setId(10L);
         shop.setAssignedManager(managerUser);
+
+        ImageInfo img = new ImageInfo();
+        img.setS3Key("shop-pic.jpg");
+        shop.setImage(img);
+
+        addressDTO = new AddressDTO();
+        addressDTO.setAlias("Main Store");
+        addressDTO.setStreet("Gran Via");
+        addressDTO.setNumber("12");
+        addressDTO.setCity("Madrid");
+        addressDTO.setLatitude(40.4168);
+        addressDTO.setLongitude(-3.7038);
+
+        shopDTO = new ShopDTO();
+        shopDTO.setName("New Shop Name");
+        shopDTO.setAssignedBudget(5000.0);
+        shopDTO.setAddress(addressDTO);
+    }
+
+    @Nested
+    @DisplayName("Tests for basic Shop creation and updates")
+    class CrudTests {
+
+        @Test
+        @DisplayName("createShop maps DTO properties to a new Entity and saves it")
+        void createShop_Success() {
+            when(shopService.save(any(Shop.class))).thenAnswer(i -> i.getArgument(0));
+
+            Shop result = orchestrator.createShop(shopDTO);
+
+            assertEquals("New Shop Name", result.getName());
+            assertEquals(5000.0, result.getAssignedBudget());
+            assertNotNull(result.getAddress());
+            assertEquals("Main Store", result.getAddress().getAlias());
+            assertEquals(40.4168, result.getAddress().getLatitude());
+
+            verify(shopService).save(any(Shop.class));
+        }
+
+        @Test
+        @DisplayName("updateShop updates existing entity fields using DTO data")
+        void updateShop_Success() {
+            when(shopService.findShopHelper(10L)).thenReturn(shop);
+
+            Shop result = orchestrator.updateShop(10L, shopDTO);
+
+            assertEquals("New Shop Name", result.getName());
+            assertEquals(5000.0, result.getAssignedBudget());
+            assertEquals("Madrid", result.getAddress().getCity());
+        }
+
+        @Test
+        @DisplayName("deleteShop unlinks memory collections, cancels active orders and deletes S3 image")
+        void deleteShop_CascadesAndUnlinks() {
+            // Setup Truck
+            Truck truck = new Truck();
+            truck.setAssignedShop(shop);
+            shop.getAssignedTrucks().add(truck);
+
+            // Setup Customer
+            User customer = new User();
+            customer.setSelectedShop(shop);
+            shop.getCustomers().add(customer);
+
+            // Setup Orders (One active, one completed)
+            Order activeOrder = new Order();
+            activeOrder.changeOrderStatus(OrderStatus.ORDER_MADE, "Init");
+            activeOrder.setAssignedShop(shop);
+
+            Order completedOrder = new Order();
+            completedOrder.changeOrderStatus(OrderStatus.COMPLETED, "Done");
+            completedOrder.setAssignedShop(shop);
+
+            shop.getAssignedOrders().addAll(List.of(activeOrder, completedOrder));
+
+            when(shopService.findShopHelper(10L)).thenReturn(shop);
+
+            try (MockedStatic<GlobalDefaults> mockedDefaults = mockStatic(GlobalDefaults.class)) {
+                mockedDefaults.when(() -> GlobalDefaults.isDefaultShopImage(shop.getImage())).thenReturn(false);
+
+                // Act
+                orchestrator.deleteShop(10L);
+
+                // Assert unlinking
+                assertTrue(shop.getAssignedTrucks().isEmpty());
+                assertNull(truck.getAssignedShop());
+
+                assertTrue(shop.getCustomers().isEmpty());
+                assertNull(customer.getSelectedShop());
+
+                assertTrue(shop.getAssignedOrders().isEmpty());
+                assertNull(activeOrder.getAssignedShop());
+                assertNull(completedOrder.getAssignedShop());
+
+                // Assert business logic: Active order must be cancelled
+                assertEquals(OrderStatus.CANCELLED, activeOrder.getCurrentStatus());
+                assertEquals("La tienda a la que estaba asignado el pedido ha sido eliminada.",
+                        activeOrder.getHistory().getLast().getUpdates().getLast().getDescription());
+
+                // Completed order must remain completed
+                assertEquals(OrderStatus.COMPLETED, completedOrder.getCurrentStatus());
+
+                // Assert deletions
+                verify(imageService).deleteFile("shop-pic.jpg");
+                verify(shopService).delete(shop);
+            }
+        }
     }
 
     @Nested
@@ -150,3 +269,4 @@ class ShopUserOrchestratorUTest {
         }
     }
 }
+
