+++ docs/pages/02-detailed-features.md
@@ -76,12 +76,12 @@
 
 ### 🔴 Advanced features
 | Ready? | Advanced feature description                              | Anon User | Registered User | Store Manager | Delivery driver | Administrator |
-|--------|-----------------------------------------------------------|-----------|-----------------|--------------|-----------------|---------------|
-| 🔄     | Check shops, trucks and/or addresses approximate location |           | ✅               | ✅            | ✅               | ✅             |
-| 🔄     | View stored use data displayed in charts                  |           |                 |              |                 | ✅             |
-| 🔄     | Import/Export Excel tables                                |           |                 | ✅            |                 | ✅             |
-| ✅     | Read unique QR code for order reception                   |           |                 | ✅            | ✅               |               |
-| ✅     | Generate unique QR code for order reception               |           | ✅               |              |                 |               |
+|--------|-----------------------------------------------------------|-----------|-----------------|-------------|-----------------|---------------|
+| 🔄     | Check shops, trucks and/or addresses approximate location |           | ✅               | ✅           | ✅               | ✅             |
+| ✅     | View stored use data displayed in charts                  |           |                 |             |                 | ✅             |
+| ✅     | Real-Time use-data tracking and history visualization     |           |                 |            |                 | ✅             |
+| ✅     | Read unique QR code for order reception                   |           |                 | ✅           | ✅               |               |
+| ✅     | Generate unique QR code for order reception               |           | ✅               |             |                 |               |
 
 &nbsp;
